requirements.txt
streamlit>=1.22.0
Pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
scikit-learn>=1.2.0 